requirements.txt
aiohttp @ file:///home/conda/feedstock_root/build_artifacts/aiohttp_1649013154501/work
aiosignal @ file:///home/conda/feedstock_root/build_artifacts/aiosignal_1667935791922/work
appdirs @ file:///home/conda/feedstock_root/build_artifacts/appdirs_1603108395799/work
async-timeout @ file:///home/conda/feedstock_root/build_artifacts/async-timeout_1640026696943/work
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1671632566681/work
Bottleneck @ file:///opt/conda/conda-bld/bottleneck_1657175564434/work
brotlipy @ file:///home/conda/feedstock_root/build_artifacts/brotlipy_1648854164373/work
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1636046055389/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1661170624537/work
click @ file:///tmp/build/80754af9/click_1646056590078/work
cmake==3.25.2
colorama @ file:///croot/colorama_1672386526460/work
cryptography @ file:///home/conda/feedstock_root/build_artifacts/cryptography_1637687020823/work
cycler @ file:///home/conda/feedstock_root/build_artifacts/cycler_1635519461629/work
docker-pycreds==0.4.0
et-xmlfile @ file:///home/conda/feedstock_root/build_artifacts/et_xmlfile_1674664118162/work
ffmpeg-python==0.2.0
filelock==3.9.0
frozenlist @ file:///croot/frozenlist_1670004507010/work
future==0.18.3
gitdb @ file:///home/conda/feedstock_root/build_artifacts/gitdb_1669279893622/work
GitPython @ file:///home/conda/feedstock_root/build_artifacts/gitpython_1676637751542/work
huggingface-hub==0.12.1
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1663625384323/work
joblib @ file:///home/conda/feedstock_root/build_artifacts/joblib_1663332044897/work
kiwisolver @ file:///croot/kiwisolver_1672387140495/work
lit==15.0.7
llvmlite==0.39.1
matplotlib @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-suite_1636786646663/work
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
more-itertools==9.1.0
multidict @ file:///croot/multidict_1665674239670/work
numba==0.56.4
numexpr @ file:///croot/numexpr_1668713893690/work
numpy @ file:///croot/numpy_and_numpy_base_1672336185480/work
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
openai @ file:///home/conda/feedstock_root/build_artifacts/openai_1677707969558/work
openai-whisper==20230307
openpyxl @ file:///home/conda/feedstock_root/build_artifacts/openpyxl_1632332968451/work
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1673482170163/work
pandas==1.5.3
pandas-stubs @ file:///home/conda/feedstock_root/build_artifacts/pandas-stubs_1677983933187/work
pathtools==0.1.2
Pillow==9.4.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1677883256757/work
plotly @ file:///home/conda/feedstock_root/build_artifacts/plotly_1677271849976/work
pooch @ file:///home/conda/feedstock_root/build_artifacts/pooch_1677600624926/work
protobuf==3.20.3
psutil @ file:///opt/conda/conda-bld/psutil_1656431268089/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1608055815057/work
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1652235407899/work
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1661604839144/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytube @ file:///home/conda/feedstock_root/build_artifacts/pytube_1671251403801/work
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1673864280276/work
PyYAML @ file:///home/conda/feedstock_root/build_artifacts/pyyaml_1648757097602/work
regex==2022.10.31
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1673863902341/work
scikit-learn @ file:///croot/scikit-learn_1676911643119/work
scipy==1.10.0
sentry-sdk @ file:///home/conda/feedstock_root/build_artifacts/sentry-sdk_1677505182752/work
setproctitle @ file:///home/conda/feedstock_root/build_artifacts/setproctitle_1649637314806/work
shellingham @ file:///croot/shellingham_1669142169426/work
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smmap @ file:///home/conda/feedstock_root/build_artifacts/smmap_1611376390914/work
tenacity @ file:///home/conda/feedstock_root/build_artifacts/tenacity_1677600641219/work
threadpoolctl @ file:///home/conda/feedstock_root/build_artifacts/threadpoolctl_1643647933166/work
tokenizers==0.13.2
torch==1.13.1
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1648827245914/work
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1677948868469/work
transformers==4.26.1
triton==2.0.0
typer @ file:///tmp/build/80754af9/typer_1637667998014/work
types-pytz @ file:///home/conda/feedstock_root/build_artifacts/types-pytz_1677064973366/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1665144421445/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1673452138552/work
wandb @ file:///home/conda/feedstock_root/build_artifacts/wandb_1678246466502/work
yarl @ file:///home/conda/feedstock_root/build_artifacts/yarl_1648966524636/work
